+++ src/compas_notebook/scene/graphobject.py
@@ -1,4 +1,5 @@
 import pythreejs as three
+from compas.colors import Color
 from compas.scene import GraphObject
 
 from compas_notebook.conversions import nodes_and_edges_to_threejs
@@ -8,6 +9,20 @@ from compas_notebook.scene import ThreeSceneObject
 
 class ThreeGraphObject(ThreeSceneObject, GraphObject):
     """Scene object for drawing graph."""
+
+    def __init__(
+        self,
+        nodesize=0.1,
+        nodecolor=Color(0.0, 0.0, 0.0),
+        edgecolor=Color(0.2, 0.2, 0.2),
+        **kwargs,
+    ):
+        super().__init__(
+            nodesize=nodesize,
+            nodecolor=nodecolor,
+            edgecolor=edgecolor,
+            **kwargs,
+        )
 
     def draw(self):
         """Draw the graph associated with the scene object.